+++ www/go/modules/community/comments/language/sl.php
@@ -1,0 +1,14 @@
+<?php
+return array (
+  'Add labels' => 'Dodajte nalepke',
+  'Attach file' => 'Priložite datoteko',
+  'Scroll to top' => 'Pomaknite se na vrh',
+  'You have an unsaved comment. Are you sure you want to discard the comment?' => 'Imate nesposoben komentar. Ali ste prepričani, da želite komentar zavreči?',
+  'Unknown user' => 'Neznani uporabnik',
+  '{author} wrote at {date}' => '{avtor} je napisal na {datum}',
+  'Edited by {author} at {date}' => 'Uredil {avtor} na {datum}',
+  'The date was changed to {date}' => 'Datum je bil spremenjen v {datum}',
+  'Labels' => 'Nalepke',
+  'Message' => 'Sporočilo',
+  'A comment was made by {creator}' => 'Komentar je podal {Creator}',
+);