+++ www/go/modules/community/imapauthenticator/language/sl.php
@@ -1,0 +1,13 @@
+<?php
+return array (
+  'Enter the domains this imap server should be used to authenticate. Users must login with their e-mail address and if the domain matches this profile it will be used.' => 'Vnesite domene Ta strežnik IMAP je treba uporabiti za preverjanje pristnosti. Uporabniki se morajo prijaviti z e-poštnim naslovom in če se domena ujema s tem profilom, bo uporabljen.',
+  'Use a \'*\' to match all domains.' => 'Uporabite \'*\' za ujemanje vseh domen.',
+  'Domains' => 'Domene',
+  'Validate certificate' => 'Preverjanje potrdila',
+  'Remove domain from username' => 'Odstranite domeno iz uporabniškega imena',
+  'Users must login with their full e-mail adress. Enable this option if the IMAP expects. the username without domain.' => 'Uporabniki se morajo prijaviti s svojim celotnim e-poštnim naslovom. Omogoči to možnost, če IMAP pričakuje. Uporabniško ime brez domene.',
+  'Use IMAP credentials' => 'Uporabite poverilnice IMAP',
+  'Enable this if the SMTP server credentials are identical to the IMAP server.' => 'Omogočite to, če so poverilnice strežnika SMTP enake strežniku IMAP.',
+  'User options' => 'Možnosti uporabnika',
+  'IMAP Authenticator' => 'IMAP AUTRATICATOR',
+);