+++ www/go/modules/community/notesencrypt/language/sl.php
@@ -1,0 +1,9 @@
+<?php
+return array (
+  'Encrypted data' => 'Šifrirani podatki',
+  'Cryptographic functions are not available. Please run Group-Office with SSL.' => 'Kriptografske funkcije niso na voljo. Prosimo, zaženite skupinsko pisarno s SSL.',
+  'Enter password to decrypt' => 'Vnesite geslo za dešifriranje',
+  'Wrong password' => 'Napačno geslo',
+  'Sorry, you can\'t use images in encrypted notes' => 'Žal ne morete uporabljati slik v šifriranih opombah',
+  'Notes encrypt' => 'Opombe šifrirajo',
+);