+++ www/modules/reminders/language/sl.php
@@ -1,0 +1,8 @@
+<?php
+return array (
+  'Popup reminder' => 'Popup opomnik',
+  'Add users' => 'Dodajte uporabnike',
+  'Add user groups' => 'Dodajte skupine uporabnikov',
+  'Snooze time' => 'Čas dremeža',
+  'Popup reminders' => 'Popup opomniki',
+);