+++ CHANGELOG.md
@@ -1,6 +1,39 @@
+- Core: Added Slovenian / Slovenščina translation
+- Postfixadmin: Added "fts" option to enable full text search
+- Email: Full text search is enabled if the mailserver returns XFTS as capability (custom capability used in Group-office mailserver) or
+  when this config option is set:
+
+'community' => [
+	'email' => [
+		'forceFTS' => [
+			'<HOSTNAME>' => true
+			]
+		]
+	]
+];
+
+23-04-2024: 6.8.44
 - Studio: fix dependency and minor deprecation errors
 - DAV core: DAV principals didn't apply user permissions
+- Core: Don't re-open tabs on badge notification change
+- Core: close tab with menu and not directly with right click
+- Email: Allow full message search by default
+- Email: Select current or all folders from menu
+- Email: New config option to define a hidden folder that shows all:
 
+'community' => [
+	'email' => [
+		'allFolder' => [
+			'<HOSTNAME>' => 'virtual/All'
+			]
+		]
+	]
+];
+
+This folder will be available in the group-office mailserver by default.
+
+- Core: Hint admins to enter their admin password when changing a user password
+- Email: minor bugfix in saved message
 
 16-04-2024: 6.8.43
 - E-mail: Fixed increasing padding when saving drafts multiple times